Expose query_all_address_maps; create enumerable.rs
