+++ contracts/cw20-merkle-airdrop/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod contract;
+mod enumerable;
 mod error;
 pub mod helpers;
 pub mod msg;
